// src/lib/api/models/BookingStatusEnum.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * * `pending` - Pending
 * * `confirmed` - Confirmed
 * * `cancelled` - Cancelled
 * * `completed` - Completed
 */
export enum BookingStatusEnum {
    PENDING = 'pending',
    CONFIRMED = 'confirmed',
    CANCELLED = 'cancelled',
    COMPLETED = 'completed',
}